Add TransportStopHelper & remove second search for amenity in wpt & favourite menu builders
